Added a new scene with rune system
Added a lot of scripts which are being used to create an inventory with moveable items that snap to the different slots. Also started the framework for scriptable items that will be useful when we try to make items that actually do stuff.  Still in the middle of that, though, so all it does is add a sprite to each item.